Add region events for layout children views
Region events are one of the most common requets we've received:
events in a layout like you can listen for child events in a collection
view.

```js
LayoutView.extend({
  onChildViewSubmit: function(childView, successMsg) {},
  onChildViewFail: function(childView, failMsg) {}
```

Workarounds:
The first is share a radio channel,
the second is share a common eventable object (think model). Messaging
through a channel can be ugly because it encourages an explosion of
channels that can be hard to keep track of. Think lots of little wires
through the app. Messaging through a shared eventable object, is ugly
because it's a source of indirection that's unexpected. It's not obvious
that the model is being used in this way.

Event Bubbling:
This is not event bubbling where an event will go all the way up the
view tree. I do not want to rule that out, as it might be nice, but
there are more design considerations there.

When we introduced the `_parent` reference, we discussed that there was
support for getting nested children for proxying events, but there was
no support for getting ancestor views for proxying events and other
internal utilities. We now have support for this, but will obviously
discourage end users from tapping into the power. With great power
comes...